Adicionando anotação de tipo e formatando código que nem gente
