+++ spy/__init__.py
@@ -1,0 +1,1 @@
+from .spy import SPY, Crawler, Parser
