+++ spy/web_parser.py
@@ -1,10 +1,13 @@
+from parsel import Selector
+
 class Parser:
     def __init__(self):
         pass
 
-    def get_items_group(self, group_selector, model, page):
-        data = page.css(group_selector)
+    def get_items_group(self, query: str, model: dict, page: Selector) -> list[dict]:
+        data = page.css(query)
         dataset = []
+
         for e in data:
             dataset.append({ k: e.css(v).get() for k,v in model.items() })
